requirements.txt
googletrans==3.0.0
openai==0.27.2
python-telegram-bot==20.3
Requests==2.31.0
